+++ library/src/methods/fallback/fallback.test.ts
@@ -1,117 +1,30 @@
-import { describe, test, expect, expectTypeOf } from 'vitest';
-import { object, string, number } from '../../schemas/index.ts';
-import { minLength } from '../../validations/index.ts';
-import { Issue, Issues, parse, safeParse } from '../../index.ts';
+import { describe, expect, test } from 'vitest';
+import { object, string } from '../../schemas/index.ts';
+import { parse } from '../parse/index.ts';
 import { fallback } from './fallback.ts';
 
 describe('fallback', () => {
-  test('should parse normally', () => {
-    const schema = fallback(string(), 'world');
-    const output = parse(schema, 'hello');
-    expect(output).toBe('hello');
-    expectTypeOf(output).toEqualTypeOf<string>();
+  const schema1 = fallback(string(), 'test');
+  const schema2 = fallback(string(), () => 'test');
+  const schema3 = object({ key1: schema1, key2: schema2 });
+
+  test('should use default value', () => {
+    const output1 = parse(schema1, 123);
+    expect(output1).toBe('test');
+    const output2 = parse(schema2, 123);
+    expect(output2).toBe('test');
+    const output3 = parse(schema3, {});
+    expect(output3).toEqual({ key1: 'test', key2: 'test' });
   });
 
-  test('should succeed to parse but return fallback value', () => {
-    const schema = fallback(string([minLength(6)]), 'hello world');
-    const output = parse(schema, 'hello');
-    expect(output).toEqual('hello world');
-  });
-
-  test('should succeed, return fallback and issues need to match non-fallback ones', () => {
-    const objectSchema = object({
-      text: string([minLength(6)]),
-      data: number(),
-    });
-
-    let issues: any;
-    const fallbackValue = { text: 'hello world', data: 5 };
-    const schema = fallback(objectSchema, fallbackValue, (dataIssues) => {
-      issues = dataIssues;
-    });
-
-    const data = { text: 'hello' };
-    const output = parse(schema, data);
-    expectTypeOf(output).toEqualTypeOf<{ text: string; data: number }>();
-    expect(output).toEqual(fallbackValue);
-
-    expect(issues).lengthOf.above(0);
-    //expect(issues).toEqual([  ]);
-
-    const result = safeParse(objectSchema, data);
-    expect(result.success).toBeFalsy();
-    expect((result as any).issues).toEqual(issues);
-  });
-
-  test('nested fallback', () => {
-    let warnings: any;
-    const fallbackValue = { text: 'hello world', data: 5 };
-    const schema = fallback(
-      object({
-        text: fallback(
-          string([minLength(6)]),
-          fallbackValue.text,
-          (dataIssues) => {
-            warnings = dataIssues;
-          }
-        ),
-        data: number(),
-      }),
-      fallbackValue,
-      (dataIssues) => {
-        warnings = dataIssues;
-      }
-    );
-
-    {
-      const data = { data: 2 };
-      const output = parse(schema, data);
-      expect(output).toEqual({ ...data, text: fallbackValue.text });
-
-      expect(warnings).lengthOf.above(0);
-      expect(warnings).toHaveLength(1);
-      //expect(issues).toEqual([  ]);
-    }
-
-    warnings = [];
-    {
-      const output = parse(schema, {});
-      expect(output).toEqual(fallbackValue);
-
-      expect(warnings).lengthOf.above(0);
-      expect(warnings).toHaveLength(1);
-    }
-  });
-
-  test('collect warnings', () => {
-    const warnings: Issue[] = [];
-    const collect = (issues: Issues) =>
-      issues.forEach((value) => warnings.push(value));
-
-    const fallbackValue = { text: 'hello world', data: 5 };
-    const schema = fallback(
-      object({
-        text: fallback(string([minLength(6)]), fallbackValue.text, collect),
-        data: number(),
-      }),
-      fallbackValue,
-      collect
-    );
-
-    {
-      const data = { data: 2 };
-      const output = parse(schema, data);
-      expect(output).toEqual({ ...data, text: fallbackValue.text });
-      console.log(warnings);
-      console.log(warnings[0].path);
-    }
-
-    {
-      const output = parse(schema, {});
-      expect(output).toEqual(fallbackValue);
-      console.log(warnings);
-    }
-
-    expect(warnings).toHaveLength(3);
+  test('should not use default value', () => {
+    const input1 = 'hello';
+    const output1 = parse(schema1, input1);
+    expect(output1).toBe(input1);
+    const output2 = parse(schema2, input1);
+    expect(output2).toBe(input1);
+    const input2 = { key1: 'hello', key2: 'hello' };
+    const output3 = parse(schema3, input2);
+    expect(output3).toEqual(input2);
   });
 });